Add files via upload
This was a project from my Operating Systems class in which I was instructed to simulate threads ( as either a man or a woman attempting to use a unisex restroom with three stalls) being restricted by locks ( as bathroom doors). The point of the project was to ensure that (male) threads wouldn't be able to access a stall if at least one (woman) thread was using the restroom. If at least one woman in the line used the restroom, to be fair, the next threads allowed in would be 1-3 (male) threads (given there were still some in line).